# docs/backend/openapi.yml
openapi: 3.1.0
info:
  title: Practice Catalog API
  description: |
    Бэкенд мобильного каталога товаров (Добрушский фарфоровый завод).

    Соглашения:
    - Деньги — целые копейки (BYN), `integer`/`int64`.
    - Даты и время — ISO-8601, UTC (`2026-08-17T12:00:00Z`).
    - Ошибки — единый формат RFC 9457 Problem Details (схема `ProblemDetails`).
    - Пагинация — `page` (0-based) + `size` (по умолчанию 20, максимум 100); конверт `{items, page, size, total, totalPages}`.
    - Идемпотентность запросов, меняющих состояние, — заголовок `X-Request-Id`.

    Подробнее: [03-api-design.md](03-api-design.md), [02-domain-model.md](02-domain-model.md).
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальная разработка
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/admin/categories:
    get:
      tags: [Admin]
      summary: Полное дерево категорий
      description: |
        Включает неактивные и мягко удалённые категории (публичное дерево отдаёт только активные).
        Узел содержит isActive, deletedAt и activeProductCount — для управления каталогом
        и поиска скрытых категорий.
      responses:
        "200":
          description: Полное дерево категорий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCategoryNode"
              examples:
                tree:
                  value:
                    - id: a1b2c3d4-1111-4c1e-9f2c-000000000001
                      name: Сервизы
                      slug: servizy
                      imageUrl: null
                      sortOrder: 1
                      isActive: false
                      deletedAt: null
                      activeProductCount: 12
                      children:
                        - id: a1b2c3d4-1111-4c1e-9f2c-000000000002
                          name: Чайные сервизы
                          slug: chainye-servizy
                          imageUrl: null
                          sortOrder: 1
                          isActive: true
                          deletedAt: null
                          activeProductCount: 5
                          children: []
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin]
      summary: Создать категорию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
            examples:
              create:
                value:
                  name: Чайные сервизы
                  slug: chainye-servizy
                  parentId: a1b2c3d4-1111-4c1e-9f2c-000000000001
                  description: Сервизы для чая
                  imageUrl: null
                  sortOrder: 1
      responses:
        "201":
          description: Категория создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryNode"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/admin/categories/{id}:
    patch:
      tags: [Admin]
      summary: Обновить категорию
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
      responses:
        "200":
          description: Обновлённая категория
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryNode"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Admin]
      summary: Удалить категорию (soft delete)
      description: Запрещено, если есть активные товары или подкатегории.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Категория удалена
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-токен, выданный через /api/v1/auth/login или /api/v1/auth/refresh
...
  responses:
    ValidationError:
      description: Ошибка валидации полей запроса
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetails"
    Unauthorized:
      description: Отсутствует, просрочен или невалиден токен
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: Недостаточно прав (роль)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Conflict:
      description: Конфликт состояния (409)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  schemas:
    ProblemDetails:
      type: object
      description: Ошибка в формате RFC 9457
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: https://api.example.com/errors/not-found
        title:
          type: string
          example: Ресурс не найден
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: "Товар с id 123 не найден"
        instance:
          type: string
          example: /api/v1/products/123
        traceId:
          type: string
          description: Сквозной идентификатор запроса (совпадает с X-Request-Id)
          example: f7a2c9d1-4b3a-4c1e-9f2c-8d1e4b3a4c1e
    ValidationProblemDetails:
      allOf:
        - $ref: "#/components/schemas/ProblemDetails"
        - type: object
          properties:
            errors:
              type: object
              description: "Карта ошибок валидации: поле → список сообщений"
              additionalProperties:
                type: array
                items:
                  type: string
              example:
                email: ["должен быть валидным email"]
                password: ["длина от 8 до 72 символов"]
...
    CategoryNode:
      type: object
      required: [id, name, slug, sortOrder, children]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 200
        imageUrl:
          type: string
          format: uri
          nullable: true
        sortOrder:
          type: integer
        children:
          type: array
          items:
            $ref: "#/components/schemas/CategoryNode"
...
    CategoryRequest:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, maxLength: 200 }
        slug: { type: string, maxLength: 200, pattern: "^[a-z0-9-]+$", description: Уникален среди активных категорий }
        parentId: { type: string, format: uuid, nullable: true }
        description: { type: string, nullable: true }
        imageUrl: { type: string, format: uri, nullable: true }
        sortOrder: { type: integer, default: 0 }
...
    AdminCategoryNode:
      type: object
      description: Узел полного дерева категорий для администратора
      required: [id, name, slug, sortOrder, isActive, activeProductCount, children]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 200 }
        slug: { type: string, maxLength: 200 }
        imageUrl: { type: string, format: uri, nullable: true }
        sortOrder: { type: integer }
        isActive:
          type: boolean
          description: false — скрыта из публичного каталога
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: Момент мягкого удаления; null — не удалена
        activeProductCount:
          type: integer
          minimum: 0
          description: Число неудалённых товаров в категории
        children:
          type: array
          items:
            $ref: "#/components/schemas/AdminCategoryNode"
...